+++ src/transform/defaultblock.js
@@ -21,6 +21,8 @@ function replaceBodyToUseDefaultBlock( node ) {
     if ( undefined === useBlockNode ) {
         const useDefaultBlockNode = new ast.UseBlockStatementNode(
             new ast.LiteralNode( '\'default\'' ),
+            new ast.IdentifierNode( 'blockData' ),
+            false,
             node.body,
             node.loc
         );